requirements.txt
psycopg2-binary
Django
dsnparse
django-webpack-loader
django-markdownx
django-markdown-deux
gunicorn
pyaml
-e git://github.com/disqus/django-bitfield.git@master#egg=django-bitfield
